pyproject.toml
[project]
name = "tender-expert"
version = "0.1.0"
description = ""
authors = [
    {name = "liinaD3210",email = "vodenikov.daniil@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.46.1,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pymupdf (>=1.26.1,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "google-api-python-client (>=2.175.0,<3.0.0)",
    "xlsxwriter (>=3.2.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
